requirements.txt
attrs==19.3.0
certifi==2020.4.5.1
chardet==3.0.4
decorator==4.4.2
future==0.18.2
huepy==1.2.1
idna==2.9
imageio==2.8.0
imageio-ffmpeg==0.4.2
instabot==0.117.0
mock==4.0.2
more-itertools==8.3.0
moviepy==1.0.3
numpy==1.18.5
packaging==20.4
Pillow==7.1.2
pluggy==0.13.1
praw==7.0.0
prawcore==1.4.0
proglog==0.1.9
py==1.8.1
pyparsing==2.4.7
PySocks==1.7.1
pytest==5.4.3
python-dotenv==0.13.0
pytz==2020.1
requests==2.23.0
requests-toolbelt==0.9.1
responses==0.10.14
schedule==0.6.0
six==1.15.0
tqdm==4.46.1
update-checker==0.17
urllib3==1.25.9
wcwidth==0.2.3
websocket-client==0.57.0
